rename checkout/ebanx.js to checkout/processor-specific.js, extract address fields validation into own object
